requirements.txt
mysql-connector-python==9.3.0
python-dotenv==1.0.0
python-dateutil==2.8.2